requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
cryptography==36.0.1
h11==0.13.0
idna==3.3
loguru==0.5.3
numpy
opencv-python==4.5.5.62
outcome==1.1.0
Pillow
pycparser==2.21
pyOpenSSL==21.0.0
python-dotenv==0.19.2
requests==2.27.0
schedule==1.1.0
selenium
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
win32-setctime==1.0.4
wsproto==1.0.0
wxpusher==2.2.0
